requirements.txt
# Core AI/LLM packages
openai>=1.3.8
langchain>=0.1.10
langchain-community>=0.1.0
langchain-text-splitters>=0.0.1
langchain-groq>=0.0.64

# Database / ORM
SQLAlchemy>=2.0.20

# Document handling
PyPDF2>=3.0.0
python-docx>=0.8.12
beautifulsoup4>=4.12.2

# Vector stores
pinecone-client>=2.2.1
faiss-cpu>=1.7.4.post2

# Web / API
fastapi>=0.111.2
uvicorn>=0.24.0
fastapi[all]
python-multipart>=0.0.6

# Utilities
pydantic>=2.4.2
typing-extensions>=4.7.1
requests>=2.31.0
python-dotenv>=1.0.0
asyncio

# Logging / debugging
python-json-logger>=2.0.7

langgraph>=0.0.64

#vision applications
open_clip_torch 
torch 
torchvision 
datasets 
accelerate 
datasets 
fsspec 
aiohttp 
huggingface_hub
clip


weaviate-client
transformers 
accelerate 
bitsandbytes 
qwen-vl-utils
gradio_client
google-generativeai